+++ app/src/main/java/com/svalero/toprestaurants/adapter/RestaurantAdapter.java
@@ -2,6 +2,7 @@ package com.svalero.toprestaurants.adapter;
 
 import static com.svalero.toprestaurants.db.Constants.DATABASE_NAME;
 
+import android.app.AlertDialog;
 import android.content.Context;
 import android.content.Intent;
 import android.view.LayoutInflater;
@@ -13,9 +14,11 @@ import android.widget.TextView;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.room.Room;
 
+import com.svalero.toprestaurants.ModifyRestaurantActivity;
 import com.svalero.toprestaurants.R;
 import com.svalero.toprestaurants.RestaurantDetailsActivity;
 import com.svalero.toprestaurants.db.AppDatabase;
+import com.svalero.toprestaurants.domain.Reserve;
 import com.svalero.toprestaurants.domain.Restaurant;
 
 import java.util.List;
@@ -53,6 +56,7 @@ public class RestaurantAdapter extends RecyclerView.Adapter<RestaurantAdapter.Re
         public TextView restaurantName;
         public TextView restaurantType;
         public Button restaurantDetailsButton;
+        public Button modifyRestaurantButton;
         public Button deleteRestaurantButton;
         public View parentView;
 
@@ -63,9 +67,11 @@ public class RestaurantAdapter extends RecyclerView.Adapter<RestaurantAdapter.Re
             restaurantName = view.findViewById(R.id.restaurant_name);
             restaurantType = view.findViewById(R.id.restaurant_type);
             restaurantDetailsButton = view.findViewById(R.id.restaurant_details_button);
+            modifyRestaurantButton = view.findViewById(R.id.modify_restaurant_button);
             deleteRestaurantButton = view.findViewById(R.id.delete_restaurant_button);
 
             restaurantDetailsButton.setOnClickListener(v -> seeDetails(getAdapterPosition()));
+            modifyRestaurantButton.setOnClickListener(v -> modifyRestaurant(getAdapterPosition()));
             deleteRestaurantButton.setOnClickListener(v -> deleteRestaurant(getAdapterPosition()));
         }
     }
@@ -78,13 +84,29 @@ public class RestaurantAdapter extends RecyclerView.Adapter<RestaurantAdapter.Re
         context.startActivity(intent);
     }
 
-    private void deleteRestaurant(int position) {
-        final AppDatabase db = Room.databaseBuilder(context, AppDatabase.class, DATABASE_NAME)
-                .allowMainThreadQueries().build();
+    private void modifyRestaurant(int position){
         Restaurant restaurant = restaurantsList.get(position);
-        db.restaurantDao().delete(restaurant);
 
-        restaurantsList.remove(position);
-        notifyItemRemoved(position);
+        Intent intent = new Intent(context, ModifyRestaurantActivity.class);
+        intent.putExtra("id", restaurant.getId());
+        context.startActivity(intent);
+    }
+
+    private void deleteRestaurant(int position) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(context);
+        builder.setMessage("Are you sure?")
+                .setTitle("Delete restaurant")
+                .setPositiveButton("Yes", (dialog, id) -> {
+                    final AppDatabase db = Room.databaseBuilder(context, AppDatabase.class, DATABASE_NAME)
+                            .allowMainThreadQueries().build();
+                    Restaurant restaurant = restaurantsList.get(position);
+                    db.restaurantDao().delete(restaurant);
+
+                    restaurantsList.remove(position);
+                    notifyItemRemoved(position);
+                })
+                .setNegativeButton("No", (dialog, id) -> dialog.dismiss());
+        AlertDialog dialog = builder.create();
+        dialog.show();
     }
 }
